requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pyTelegramBotAPI==4.13.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
urllib3==2.0.4
